Rename entry_point to main
